requirements.txt
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
fastapi==0.53.2
google-api-core==1.16.0
google-auth==1.12.0
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
h11==0.9.0
# httptools==0.1.1
idna==2.9
importlib-metadata==1.6.0
more-itertools==8.2.0
# numpy==1.18.3
packaging==20.3
# 

# pluggy==0.13.1
# protobuf==3.11.3
# py==1.8.1
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
pydantic==1.4
# pyparsing==2.4.6
# pytest==5.4.1
# python-dateutil==2.8.1
# pytz==2019.3
requests==2.23.0
# rsa==4.0
# ruamel.yaml==0.16.10
# ruamel.yaml.clib==0.2.0
# six==1.14.0
# SQLAlchemy==1.3.15
starlette==0.13.2
ujson==2.0.3
# urllib3==1.25.8
uvicorn==0.11.3
# uvloop==0.14.0
wcwidth==0.1.9
websockets==8.1
zipp==3.1.0
